INI 52 <- M

+++ app/Actions/Catalogue/ProductCategory/UI/IndexFamilies.php
@@ -131,7 +131,6 @@ class IndexFamilies extends OrgAction
             } elseif ($parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
                 $queryBuilder->where('product_categories.sub_department_id', $parent->id);
             } else {
-                // todo
                 abort(419);
             }
         }
@@ -162,7 +161,8 @@ class IndexFamilies extends OrgAction
                 'organisations.slug as organisation_slug',
                 'product_category_sales_intervals.sales_grp_currency_all as sales_all',
                 'product_category_ordering_intervals.invoices_all as invoices_all',
-                DB::raw("(
+                DB::raw(
+                    "(
                     SELECT json_agg(json_build_object(
                         'id', c.id,
                         'slug', c.slug,
@@ -174,7 +174,8 @@ class IndexFamilies extends OrgAction
                     WHERE chm.model_id = product_categories.id
                         AND chm.model_type = 'ProductCategory'
                         AND c.deleted_at IS NULL
-                )::text as collections"),
+                )::text as collections"
+                ),
             ])
             ->leftJoin('product_category_stats', 'product_categories.id', 'product_category_stats.product_category_id')
             ->where('product_categories.type', ProductCategoryTypeEnum::FAMILY)
@@ -188,23 +189,28 @@ class IndexFamilies extends OrgAction
                 'number_current_products',
                 'sub_department_name',
                 'department_name',
-                AllowedSort::custom('collections', new class implements Sort {
-                    public function __invoke(Builder $query, bool $descending, string $property)
-                    {
-                        $direction = $descending ? 'desc' : 'asc';
-                        $query->orderBy(
-                            DB::raw("(
+                AllowedSort::custom(
+                    'collections',
+                    new class () implements Sort {
+                        public function __invoke(Builder $query, bool $descending, string $property)
+                        {
+                            $direction = $descending ? 'desc' : 'asc';
+                            $query->orderBy(
+                                DB::raw(
+                                    "(
                                 SELECT json_agg(c.name)
                                 FROM collection_has_models chm
                                 JOIN collections c ON chm.collection_id = c.id
                                 WHERE chm.model_id = product_categories.id
                                 AND chm.model_type = 'ProductCategory'
                                 AND c.deleted_at IS NULL
-                            )::text"),
-                            $direction
-                        );
+                            )::text"
+                                ),
+                                $direction
+                            );
+                        }
                     }
-                })
+                )
             ])
             ->allowedFilters([$globalSearch])
             ->withPaginator($prefix, tableName: request()->route()->getName())
@@ -217,7 +223,7 @@ class IndexFamilies extends OrgAction
             if ($prefix) {
                 $table
                     ->name($prefix)
-                    ->pageName($prefix . 'Page');
+                    ->pageName($prefix.'Page');
             }
 
             foreach ($this->getElementGroups($parent) as $key => $elementGroup) {
@@ -307,25 +313,23 @@ class IndexFamilies extends OrgAction
     public function getActions(ActionRequest $request): array
     {
         $actions = [];
-        if ($this->canEdit) {
-            if ($this->parent instanceof ProductCategory) {
-                $createRoute = "grp.org.shops.show.catalogue.departments.show.families.create";
+        if ($this->canEdit && $this->parent instanceof ProductCategory) {
+            $createRoute = "grp.org.shops.show.catalogue.departments.show.families.create";
 
-                if ($this->parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
-                    $createRoute = "grp.org.shops.show.catalogue.departments.show.sub_departments.show.family.create";
-                }
-
-                $actions[] = [
-                    'type'    => 'button',
-                    'style'   => 'create',
-                    'tooltip' => __('new family'),
-                    'label'   => __('family'),
-                    'route'   => [
-                        'name'       => $createRoute,
-                        'parameters' => $request->route()->originalParameters()
-                    ]
-                ];
+            if ($this->parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
+                $createRoute = "grp.org.shops.show.catalogue.departments.show.sub_departments.show.family.create";
             }
+
+            $actions[] = [
+                'type'    => 'button',
+                'style'   => 'create',
+                'tooltip' => __('new family'),
+                'label'   => __('family'),
+                'route'   => [
+                    'name'       => $createRoute,
+                    'parameters' => $request->route()->originalParameters()
+                ]
+            ];
         }
 
 
@@ -390,30 +394,28 @@ class IndexFamilies extends OrgAction
 
         $routes = null;
 
-        if ($this->parent instanceof PRoductCategory) {
-            if ($this->parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
-                $routes = [
-                    'attach' => [
-                        'name'       => 'grp.models.sub-department.families.attach',
-                        'parameters' => [
-                            'subDepartment' => $this->parent->id
-                        ]
-                    ],
-                    'detach' => [
-                        'method'  => 'delete',
-                        'name'       => 'grp.models.sub-department.family.detach',
-                        'parameters' => [
-                            'subDepartment' => $this->parent->id
-                        ]
-                    ],
-                    'fetch_families'   => [
-                        'name'      =>  'grp.json.product_category.families.index',
-                        'parameters' => [
-                            'productCategory' => $this->parent->slug
-                        ]
-                    ],
-                ];
-            }
+        if ($this->parent instanceof ProductCategory && $this->parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
+            $routes = [
+                'attach'         => [
+                    'name'       => 'grp.models.sub-department.families.attach',
+                    'parameters' => [
+                        'subDepartment' => $this->parent->id
+                    ]
+                ],
+                'detach'         => [
+                    'method'     => 'delete',
+                    'name'       => 'grp.models.sub-department.family.detach',
+                    'parameters' => [
+                        'subDepartment' => $this->parent->id
+                    ]
+                ],
+                'fetch_families' => [
+                    'name'       => 'grp.json.product_category.families.index',
+                    'parameters' => [
+                        'productCategory' => $this->parent->slug
+                    ]
+                ],
+            ];
         }
